+++ src/app/auth/login/constants/login.constants.ts
@@ -1,0 +1,4 @@
+export const SESSION_TYPES = Object.freeze({
+  LOGGED: 'Success',
+  INVALID_CREDENTIALS: 'Invalid credentials',
+})