+++ src/components/ui/alert/constants/alert.constants.ts
@@ -1,0 +1,4 @@
+import type { AlertType } from '@/interfaces'
+
+export const DANGER_ALERT_TYPE: AlertType = 'danger'
+export const SUCCESS_ALERT_TYPE: AlertType = 'success'