+++ src/constants/index.tsx
@@ -1,5 +1,0 @@
-export * from './ui/sidebar.constants'
-export * from './ui/alert.constants'
-export * from './ui/toast.constants'
-export * from './cart/cart.constants'
-export * from './gender/gender.constants'