+++ src/constants/ui/alert.constants.ts
@@ -1,4 +1,0 @@
-import type { AlertMessage, AlertType } from '@/interfaces'
-
-export const DANGER_ALERT_MESSAGE: AlertMessage = 'You must select a size*'
-export const DANGER_ALERT_TYPE: AlertType = 'danger'